Latex error fixes. Picture and paragraph modifications and more
Some more effort is put into reducing the errors and warnings from
compilation.

What took most time was to replace the eepic picture inclusion with a
pdf + latex construction. Check the *makfile* diff for details. The
inclusion procedure for the diagram included in *020.game.tex* is thus
modified.

Some lenght parameter of some macros for strategy descriptions are
ajusted for better tollerance. Ans in *010.introduction.tex* some
macros were used with wrong syntax.

Caption placement within the tabel environment is corrected in order
for lables to work. Lables and references are corrected throughtou the
document.

The propper way of achieving Scandinavian paragraph layout is to use
the *parskip*. That package is included, and all occurrencies of the
`\mypar` macro (where it make sense) are removed.